+++ database/migrations/16-v_deviation_lifesaving_gavle.sql
@@ -30,8 +30,8 @@ ALTER TABLE upstream.mv_match_lifesaving_gavle OWNER TO app;
 
 CREATE OR REPLACE VIEW upstream.v_deviation_lifesaving_gavle AS
 	SELECT
-		8 AS dataset_id,
-		16 AS layer_id,
+		8::bigint AS dataset_id,
+		16::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
